:recycle: Refactor background tasks and remove unused classes
