+++ src/mnt/ode/viewmodels/main_window.py
@@ -6,180 +6,28 @@ import logging
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from PyQt6.QtCore import QObject, QRunnable, QThreadPool, pyqtSignal, pyqtSlot
+from PyQt6.QtCore import QObject, QThreadPool, pyqtSignal, pyqtSlot
 
 from ..models import (
     ApplicationSettingsModel,
     InputSignalEncoding,
     LayoutModel,
-    LayoutVisualizationOptions,
     OperationalDomainPlotOptions,
 )
-from ..services import (
-    LayoutLoadError,
-    LayoutVisualizationError,
-    LayoutVisualizationService,
-    PixmapConversionService,
-    SQDFileService,
-)
+from ..tasks import GenerateLayoutPlotsTask, LoadFileTask, PixmapConversionTask
 from ..views.theme import get_theme_colors
 from .operational_domain import OperationalDomainViewModel
 from .settings import SettingsViewModel
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     from PyQt6.QtGui import QPixmap
 
+    from ..services import (
+        LayoutVisualizationService,
+        SQDFileService,
+    )
+
 logger = logging.getLogger(__name__)
-
-
-# --- QRunnable Task for Background File Loading ---
-class LoadFileTask(QRunnable):  # type: ignore[misc]
-    """Task to load an SQD file in a background thread."""
-
-    class Signals(QObject):  # type: ignore[misc]
-        """Signals available from a running worker thread."""
-
-        finished = pyqtSignal(LayoutModel, str, str)
-
-    def __init__(self, sqd_file_service: SQDFileService, file_path: str) -> None:
-        """Initialize the LoadFileTask.
-
-        Args:
-            sqd_file_service: Service instance for file loading.
-            file_path: Path to the SQD file.
-        """
-        super().__init__()
-        self.sqd_file_service = sqd_file_service
-        self.file_path = file_path
-        self.signals = LoadFileTask.Signals()
-
-    @pyqtSlot()  # type: ignore[misc]
-    def run(self) -> None:
-        """Execute the file loading task."""
-        layout_model_result: LayoutModel | None = None
-        error_message_result: str | None = None
-        try:
-            logger.info("LoadFileTask: Starting to load %s", self.file_path)
-            layout_model_result = self.sqd_file_service.load_layout(Path(self.file_path))
-            logger.info("LoadFileTask: Successfully loaded %s", self.file_path)
-        except LayoutLoadError as e:
-            logger.exception("LoadFileTask: LayoutLoadError for %s", self.file_path)
-            error_message_result = str(e)
-        except Exception:
-            logger.exception("LoadFileTask: Unexpected error loading %s", self.file_path)
-            error_message_result = "An unexpected error occurred during file loading."
-        finally:
-            self.signals.finished.emit(layout_model_result, self.file_path, error_message_result)
-
-
-# --- QRunnable Task for Generating Layout Plots ---
-class GenerateLayoutPlotsTask(QRunnable):  # type: ignore[misc]
-    """Task to generate layout plots in a background thread."""
-
-    class Signals(QObject):  # type: ignore[misc]
-        """Signals for this task."""
-
-        finished = pyqtSignal(list, list, str)
-
-    def __init__(
-        self,
-        layout_viz_service: LayoutVisualizationService,
-        layout_model: LayoutModel,
-        options: LayoutVisualizationOptions | None = None,
-        thread_pool: QThreadPool | None = None,
-    ) -> None:
-        """Initialize the GenerateLayoutPlotsTask.
-
-        Args:
-            layout_viz_service: Service for generating layout visualizations.
-            layout_model: The loaded layout model.
-            options: Visualization options.
-            thread_pool: Thread pool for parallel execution within the service.
-        """
-        super().__init__()
-        self.layout_viz_service = layout_viz_service
-        self.layout_model = layout_model
-        self.options = options or LayoutVisualizationOptions()
-        self.signals = GenerateLayoutPlotsTask.Signals()
-        self.thread_pool = thread_pool
-
-    @pyqtSlot()  # type: ignore[misc]
-    def run(self) -> None:
-        """Execute the plot generation task.
-
-        Raises:
-            LayoutVisualizationError: Not actually passed through, but caught internally.
-        """
-        distance_svgs: list[bytes | None] = []
-        presence_svgs: list[bytes | None] = []
-        error_message: str = ""
-        try:
-            if self.layout_model.sidb_layout is None:
-                msg = "SiDB layout is missing in LayoutModel."
-                raise LayoutVisualizationError(msg)  # noqa: TRY301 - raising here is clear enough
-
-            logger.info("GenerateLayoutPlotsTask: Generating SVGs for distance-encoded layouts...")
-            distance_svgs = self.layout_viz_service.create_layout_svgs(
-                layout=self.layout_model,
-                bdl_encoding=InputSignalEncoding.DISTANCE,
-                options=self.options,
-                thread_pool=self.thread_pool,
-            )
-
-            logger.info("GenerateLayoutPlotsTask: Generating SVGs for presence-encoded layouts...")
-            presence_svgs = self.layout_viz_service.create_layout_svgs(
-                layout=self.layout_model,
-                bdl_encoding=InputSignalEncoding.PRESENCE,
-                options=self.options,
-                thread_pool=self.thread_pool,
-            )
-            logger.info("GenerateLayoutPlotsTask: SVG generation finished.")
-
-        except LayoutVisualizationError as e:
-            logger.exception("GenerateLayoutPlotsTask: LayoutVisualizationError occurred.")
-            error_message = str(e)
-        except Exception:
-            logger.exception("GenerateLayoutPlotsTask: Unexpected error during SVG generation.")
-            error_message = "An unexpected error occurred during plot generation."
-        finally:
-            self.signals.finished.emit(distance_svgs, presence_svgs, error_message)
-
-
-class PixmapConversionTask(QRunnable):  # type: ignore[misc]
-    """QRunnable to convert SVGs to pixmaps using the pixmap_conversion_service."""
-
-    class Signals(QObject):  # type: ignore[misc]
-        """Signals for this task."""
-
-        conversion_pair_finished = pyqtSignal(list, list)  # distance_pixmaps, presence_pixmaps
-        conversion_single_finished = pyqtSignal(list)  # pixmaps for a single set
-
-    def __init__(
-        self,
-        svgs1: Sequence[bytes | None],
-        svgs2: Sequence[bytes | None] | None = None,
-    ) -> None:
-        """Initialize the PixmapConversionTask.
-
-        Args:
-            svgs1: Primary list of SVGs to convert.
-            svgs2: Optional secondary list of SVGs to convert.
-        """
-        super().__init__()
-        self.svgs1 = svgs1
-        self.svgs2 = svgs2
-        self.signals = PixmapConversionTask.Signals()
-
-    def run(self) -> None:
-        """Execute the pixmap conversion task."""
-        pixmaps1 = PixmapConversionService.convert(self.svgs1)
-        if self.svgs2 is not None:
-            pixmaps2 = PixmapConversionService.convert(self.svgs2)
-            self.signals.conversion_pair_finished.emit(pixmaps1, pixmaps2)
-        else:
-            self.signals.conversion_single_finished.emit(pixmaps1)
 
 
 class MainWindowViewModel(QObject):  # type: ignore[misc]
